+++ Watchman.Discord.UnitTests/AntiSpam/SpamPunishmentStrategyTests.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using NUnit.Framework;
+
+namespace Watchman.Discord.UnitTests.AntiSpam
+{
+    [TestFixture]
+    internal class SpamPunishmentStrategyTests : AntiSpamTestsBase
+    {
+
+    }
+}
